Simplified firmware upgrade sample
