+++ examples/2.device.firmware_update/CMakeLists.txt
