+++ examples/CMakeLists.txt
@@ -22,9 +22,9 @@ add_subdirectory(c_examples)
 add_subdirectory(0.basic.enumerate)
 add_subdirectory(1.stream.imu)
 add_subdirectory(2.device.control)
+add_subdirectory(2.device.firmware_update)
 add_subdirectory(3.advanced.point_cloud)
 add_subdirectory(3.advanced.preset)
-add_subdirectory(3.advanced.device_firmware_update)
 
 # cpp examples with OpenCV Requirements
 if(${OpenCV_FOUND})
